requirements.txt
streamlit==1.43.2
beautifulsoup4==4.13.3
requests==2.31.0
pillow==10.2.0
opencv-python-headless==4.11.0.86
pytesseract==0.3.13
pandas==2.2.0
openpyxl==3.1.5
langdetect==1.0.9
matplotlib==3.8.2
zipfile36==0.1.3
tqdm==4.67.1
piexif==1.1.3
